# src/main/resources/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Blockout Mobile Gateway API
  version: 1.0.0
  description: Public and secure endpoints for the mobile app.

servers:
  - url: https://api.blockoutproject.com
  - url: http://localhost:8089

paths:
  /api/v1/mobile/public/matches/{id}:
    get:
      tags:
        - Public
      summary: Get enriched match
      description: Returns an enriched match.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichedMatch' }

  /api/v1/mobile/public/enriched-pool/{id}:
    get:
      tags:
        - Public
      summary: Get enriched pool
      description: Returns an enriched pool.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichedPool' }

  /api/v1/mobile/public/enriched-team/{id}:
    get:
      tags:
        - Public
      summary: Get enriched team
      description: Returns an enriched team.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichedTeam' }

  /api/v1/mobile/public/ffvb/pdf/{token}:
    get:
      tags:
        - Public
      summary: Get FFVB PDF
      description: Returns a PDF document.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  /api/v1/mobile/public/matches:
    get:
      tags:
        - Public
      summary: List matches
      description: Returns a paginated list of matches.
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 3 }
        - name: pool_ids
          in: query
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
        - name: team_ids
          in: query
          required: false
          schema:
            type: array
            items: { type: integer, format: int64 }
        - name: status
          in: query
          required: true
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichedDayPage' }
        204:
          description: No Content

  /api/v1/mobile/public/pools/by-ids:
    get:
      tags:
        - Public
      summary: List pools by IDs
      description: Returns pools by IDs.
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: array
            items: { type: integer, format: int64 }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PoolSummary' }

  /api/v1/mobile/public/teams/by-club/{clubId}:
    get:
      tags:
        - Public
      summary: List teams by club
      description: Returns teams for a club.
      parameters:
        - name: clubId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TeamSummary' }

  /api/v1/mobile/public/teams/by-ids:
    get:
      tags:
        - Public
      summary: List teams by IDs
      description: Returns teams by IDs.
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: array
            items: { type: integer, format: int64 }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TeamSummary' }

  /api/v1/mobile/secure/enriched-notifications:
    get:
      tags:
        - Secure
      summary: List enriched notifications
      description: Returns user notifications.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 20 }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichedUserNotificationPage' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Gender: { type: string }
    Format: { type: string }

    Club:
      type: object
      properties:
        id: { type: string }
        raw_name: { type: string }
        name: { type: string }
        city: { type: string }
        postal_code: { type: string }
        email: { type: string }
        phone_number: { type: string }
        website: { type: string }
        logo_url: { type: string }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_update: { type: string, format: date-time }

    Division:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        main_color: { type: string }
        first_gradient_color: { type: string }
        second_gradient_color: { type: string }
        third_gradient_color: { type: string }
        logo_url: { type: string }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_update: { type: string, format: date-time }

    Team:
      type: object
      properties:
        id: { type: integer, format: int64 }
        club_id: { type: string }
        raw_name: { type: string }
        name: { type: string }
        short_name: { type: string }
        league_code: { type: string }
        division_id: { type: integer, format: int64 }
        format: { $ref: '#/components/schemas/Format' }
        gender: { $ref: '#/components/schemas/Gender' }
        season: { type: string }
        followers_count: { type: integer, format: int64 }
        logo_url: { type: string }
        active: { type: boolean }
        created_at: { type: string }
        last_update: { type: string }

    TeamWithStats:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        short_name: { type: string }
        logo_url: { type: string }
        points: { type: integer }
        played: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        pointsPenalty: { type: integer }
        coefSets: { type: number, format: double }
        coefPoints: { type: number, format: double }

    TeamSummary:
      type: object
      additionalProperties: true

    EnrichedTeam:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        short_name: { type: string }
        format: { $ref: '#/components/schemas/Format' }
        gender: { $ref: '#/components/schemas/Gender' }
        season: { type: string }
        followers_count: { type: integer, format: int64 }
        division: { $ref: '#/components/schemas/Division' }
        club: { $ref: '#/components/schemas/Club' }
        pools:
          type: array
          items: { $ref: '#/components/schemas/EnrichedPool' }

    Pool:
      type: object
      properties:
        id: { type: integer, format: int64 }
        pool_code: { type: string }
        league_code: { type: string }
        season: { type: string }
        league_name: { type: string }
        raw_name: { type: string }
        name: { type: string }
        short_name: { type: string }
        division_id: { type: integer, format: int64 }
        format: { $ref: '#/components/schemas/Format' }
        gender: { $ref: '#/components/schemas/Gender' }
        followers_count: { type: integer, format: int64 }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_update: { type: string, format: date-time }

    PoolSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        short_name: { type: string }
        league_name: { type: string }
        season: { type: string }
        gender: { $ref: '#/components/schemas/Gender' }
        format: { $ref: '#/components/schemas/Format' }
        division: { $ref: '#/components/schemas/Division' }

    EnrichedPool:
      type: object
      properties:
        id: { type: integer, format: int64 }
        season: { type: string }
        pool_code: { type: string }
        league_code: { type: string }
        league_name: { type: string }
        name: { type: string }
        short_name: { type: string }
        format: { $ref: '#/components/schemas/Format' }
        gender: { $ref: '#/components/schemas/Gender' }
        followers_count: { type: integer, format: int64 }
        ranking:
          type: array
          items: { $ref: '#/components/schemas/TeamWithStats' }
        division: { $ref: '#/components/schemas/Division' }

    Match:
      type: object
      properties:
        id: { type: integer, format: int64 }
        match_code: { type: string }
        league_code: { type: string }
        pool_id: { type: integer, format: int64 }
        live_code: { type: integer, format: int64 }
        team_id_a: { type: integer, format: int64 }
        team_id_b: { type: integer, format: int64 }
        match_date: { type: string }
        season: { type: string }
        set: { type: string }
        score: { type: string }
        status: { type: string }
        venue: { type: string }
        first_referee: { type: string }
        second_referee: { type: string }
        active: { type: boolean }
        created_at: { type: string }
        last_update: { type: string }

    EnrichedMatch:
      type: object
      properties:
        id: { type: integer, format: int64 }
        live_code: { type: integer, format: int64 }
        match_date: { type: string }
        season: { type: string }
        set: { type: string }
        score: { type: string }
        status: { type: string }
        venue: { type: string }
        first_referee: { type: string }
        second_referee: { type: string }
        team_a: { $ref: '#/components/schemas/Team' }
        team_b: { $ref: '#/components/schemas/Team' }
        match_address_pdf_url: { type: string }
        match_sheet_pdf_url: { type: string }
        pool: { $ref: '#/components/schemas/EnrichedPool' }

    PoolMatches:
      type: object
      properties:
        pool_id: { type: integer, format: int64 }
        matches:
          type: array
          items: { $ref: '#/components/schemas/Match' }

    DayMatches:
      type: object
      properties:
        date: { type: string }
        pools:
          type: array
          items: { $ref: '#/components/schemas/PoolMatches' }

    DayPage:
      type: object
      properties:
        day_matches:
          type: array
          items: { $ref: '#/components/schemas/DayMatches' }
        has_next: { type: boolean }
        next_page: { type: integer }

    EnrichedPoolMatches:
      type: object
      properties:
        pool: { $ref: '#/components/schemas/EnrichedPool' }
        matches:
          type: array
          items: { $ref: '#/components/schemas/EnrichedMatch' }

    EnrichedDayMatches:
      type: object
      properties:
        date: { type: string }
        pools:
          type: array
          items: { $ref: '#/components/schemas/EnrichedPoolMatches' }

    EnrichedDayPage:
      type: object
      properties:
        day_matches:
          type: array
          items: { $ref: '#/components/schemas/EnrichedDayMatches' }
        has_next: { type: boolean }
        next_page: { type: integer }

    CompetitionAssociation:
      type: object
      properties:
        id: { type: integer, format: int64 }
        pool_id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        club_id: { type: string }
        active: { type: boolean }
        points: { type: integer }
        played: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        wins_three_to_zero: { type: integer }
        wins_three_to_one: { type: integer }
        wins_three_to_two: { type: integer }
        losses_zero_to_three: { type: integer }
        losses_one_to_three: { type: integer }
        losses_two_to_three: { type: integer }
        won_sets: { type: integer }
        lost_sets: { type: integer }
        won_points: { type: integer }
        lost_points: { type: integer }
        points_penalty: { type: integer }
        coef_sets: { type: number, format: double }
        coef_points: { type: number, format: double }
        created_at: { type: string, format: date-time }
        last_update: { type: string, format: date-time }

    TeamRanking:
      type: object
      properties:
        team_id: { type: integer, format: int64 }
        points: { type: integer }
        points_penalty: { type: integer }
        played: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        coef_sets: { type: number, format: double }
        coef_points: { type: number, format: double }

    PoolWithRanking:
      type: object
      properties:
        pool_id: { type: integer, format: int64 }
        ranking:
          type: array
          items: { $ref: '#/components/schemas/TeamRanking' }

    EnrichedUserNotificationPage:
      type: object
      properties:
        notifications:
          type: array
          items: { $ref: '#/components/schemas/EnrichedUserNotification' }
        has_next: { type: boolean }
        next_page: { type: integer }

    EnrichedUserNotification:
      type: object
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        type: { type: string }
        title: { type: string }
        body: { type: string }
        deep_link: { type: string }
        target_type: { type: string }
        target_id: { type: integer, format: int64 }
        metadata:
          type: object
          additionalProperties: true
        is_read: { type: boolean }
        is_opened: { type: boolean }
        created_at: { type: string }
        read_at: { type: string }
        opened_at: { type: string }
        division_logo_url: { type: string }